feat: enhance deployment functions
