+++ backend/tests/client.go
@@ -18,7 +18,7 @@ import (
 const (
 	clusterName    = "jrk8s01"
 	leaderNodeName = "leader"
-	workerNodeName = "worker"
+	workerNodeName = "worker2"
 	masterNodeName = "master"
 )
 
@@ -133,7 +133,7 @@ func (c *Client) DeployCluster(clusterName string) (string, error) {
 				EnvVars: map[string]string{
 					"SSH_KEY": "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQDJ1t4Ug8EfykmJwAbYudyYYN/f7dZaVg3KGD2Pz0bd9pajAAASWYrss3h2ctCZWluM6KAt289RMNzxlNUkOMJ9WhCIxqDAwtg05h/J27qlaGCPP8BCEITwqNKsLwzmMZY1UFc+sSUyjd35d3kjtv+rzo2meaReZnUFNPisvxGoygftAE6unqNa7TKonVDS1YXzbpT8XdtCV1Y6ACx+3a82mFR07zgmY4BVOixNBy2Lzpq9KiZTz91Bmjg8dy4xUyWLiTmnye51hEBgUzPprjffZByYSb2Ag9hpNE1AdCGCli/0TbEwFn9iEroh/xmtvZRpux+L0OmO93z5Sz+RLiYXKiYVV5R5XYP8y5eYi48RY2qr82sUl5+WnKhI8nhzayO9yjPEp3aTvR1FdDDj5ocB7qKi47R8FXIuwzZf+kJ7ZYmMSG7N21zDIJrz6JGy9KMi7nX1sqy7NSqX3juAasIjx0IJsE8zv9qokZ83hgcDmTJjnI+YXimelhcHn4M52hU= omar@jarvis",
 				},
-				NodeID: 152,
+				NodeID: 150,
 			},
 			{
 				Name:     masterNodeName,
@@ -145,7 +145,7 @@ func (c *Client) DeployCluster(clusterName string) (string, error) {
 				EnvVars: map[string]string{
 					"SSH_KEY": "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQDJ1t4Ug8EfykmJwAbYudyYYN/f7dZaVg3KGD2Pz0bd9pajAAASWYrss3h2ctCZWluM6KAt289RMNzxlNUkOMJ9WhCIxqDAwtg05h/J27qlaGCPP8BCEITwqNKsLwzmMZY1UFc+sSUyjd35d3kjtv+rzo2meaReZnUFNPisvxGoygftAE6unqNa7TKonVDS1YXzbpT8XdtCV1Y6ACx+3a82mFR07zgmY4BVOixNBy2Lzpq9KiZTz91Bmjg8dy4xUyWLiTmnye51hEBgUzPprjffZByYSb2Ag9hpNE1AdCGCli/0TbEwFn9iEroh/xmtvZRpux+L0OmO93z5Sz+RLiYXKiYVV5R5XYP8y5eYi48RY2qr82sUl5+WnKhI8nhzayO9yjPEp3aTvR1FdDDj5ocB7qKi47R8FXIuwzZf+kJ7ZYmMSG7N21zDIJrz6JGy9KMi7nX1sqy7NSqX3juAasIjx0IJsE8zv9qokZ83hgcDmTJjnI+YXimelhcHn4M52hU= omar@jarvis",
 				},
-				NodeID: 152,
+				NodeID: 150,
 			},
 			{
 				Name:     workerNodeName,
@@ -173,7 +173,7 @@ func (c *Client) DeployCluster(clusterName string) (string, error) {
 		return "", fmt.Errorf("deploy request failed with status %d: %s", resp.StatusCode, string(body))
 	}
 
-	var deployResp app.DeployResponse
+	var deployResp app.Response
 	if err := json.NewDecoder(resp.Body).Decode(&deployResp); err != nil {
 		return "", err
 	}
